requirements.txt
beautifulsoup4
bing-image-downloader
black
cachetools
certifi
chardet
click
colorama
configparser
crayons
decorator
pyttsx3
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-texttospeech
grpcio
gTTS
httplib2
idna
imageio
imageio-ffmpeg
libcst
moviepy
mutagen
numpy
pathspec
Pillow
proglog
protobuf
pydub
pytz
PyYAML
regex
requests
retry
selenium
six
soupsieve
termcolor
toml
tqdm
uritemplate
urllib3
webdriver-manager
